+++ lib/product_matching_v2/domain/product_match_reason.dart
@@ -6,4 +6,8 @@ enum ProductMatchReason {
   multipleCandidates,
   lowConfidence,
   noCandidate,
+  clearWinner,
+  tie,
+  insufficientEvidence,
+  noCandidates,
 }